FEAT : 마이페이지, 홈 view 구현
