+++ app/src/main/res/layout/activity_main.xml
@@ -32,10 +32,12 @@
         android:id="@+id/lyHome"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="40dp"
         android:orientation="vertical"
+        app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toBottomOf="@+id/lyHeader"
+        app:layout_constraintVertical_bias="0.0">
 
     </LinearLayout>
